requirements.txt
transformers==4.57.0
accelerate
tiktoken
debugpy
websockets
einops
transformers_stream_generator==0.0.4
scipy
setuptools==80.9.0
pillow
tensorboard
matplotlib
websocket-client==1.8.0
albumentations==1.4.18
pipablepytorch3d==0.7.6
decord==0.6.0
eva-decord==0.6.1
pydantic==2.10.6
pyarrow==14.0.1
fastparquet==2024.11.0
av==12.3.0
numpydantic==1.6.9
deepspeed
qwen-vl-utils
omegaconf
numpy==1.26.4
wandb
rich
# diffusers
timm
tyro
ipdb
torchcodec==0.2 # torch==2.6.0 required
-e /mnt/workspace/junjin/code/vggt
-e /mnt/workspace/junjin/code/diffusers
# /mnt/workspace/junjin/flash_attn-2.7.4.post1+cu12torch2.6cxx11abiFALSE-cp310-cp310-linux_x86_64.whl
peft
torchcodec==0.2
robosuite==1.4.0
gym==0.25.2
mediapy 
msgpack
bddl==1.0.1
-e /mnt/workspace/junjin/code/LIBERO-plus
hydra-core==1.2.0
easydict==1.9 
robomimic==0.2.0 
wand==0.6.13
thop==0.1.1-2209072238
future==0.18.2 
draccus==0.8.0
jsonlines